+++ TicketToCode.Core/Enums/Category.cs
@@ -1,5 +1,4 @@
 using System.ComponentModel;
-using System.Reflection;
 
 namespace TicketToCode.Core.Enums
 {
@@ -42,15 +41,5 @@ namespace TicketToCode.Core.Enums
         SlowCooker
     }
 
-    public static class EnumHelper
-    {
-        public static string GetEnumDescription(Enum value)
-        {
-            FieldInfo fi = value.GetType().GetField(value.ToString());
-            DescriptionAttribute[] attributes = (DescriptionAttribute[])fi.GetCustomAttributes(typeof(DescriptionAttribute), false);
-
-            string description = attributes != null && attributes.Length > 0 ? attributes[0].Description : value.ToString();
-            return $"{value.ToString()}: {description}";
-        }
-    }
+   
 }
